requirements.txt
Flask>=2.0
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
email-validator
